+++ tests/2xx/219_cleanup.sh
@@ -3,4 +3,3 @@
 . common.sh
 
 # no test yet
-