+++ tests/2xx/222_get_help_text.sh
@@ -2,7 +2,6 @@
 
 . common.sh
 
-#set -x
 assert_func get_help_text $TEST_FAIL $TEST_EMPTY
 CSL_HELP_TEXT='bla'
 assert_func get_help_text $TEST_OK 'bla'