+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 # 1.1
 
-* new functions set_help_text(), get_help_text() and has_help_text()
+* new functions (set|get|has)_help_text()
+* new functions add_param(), (get|has)_short_params(), (get|has)_long_params()
+* new function is_declared_func()
 
 # 1.0